Separate test cases
Not every test requires the integration test environment
